BHL/services/transcription/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
pymongo==4.6.0
celery==5.3.6
redis==5.0.1
openai-whisper==20231117
httpx==0.25.2
python-multipart==0.0.6
pydantic-settings==2.1.0